* Separate Batch class * Formatted code
