+++ molgenis-compute-core/src/main/java/org/molgenis/compute/generators/impl/BatchAnalyser.java
@@ -4,105 +4,80 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-/**
- * Created by georgebyelas on 16/03/15.
- */
+import org.molgenis.compute.model.impl.Batch;
+
 public class BatchAnalyser
 {
+	private final String batchVariable;
+	private final int batchSize;
 
-    private final String batchVariable;
-    private final int batchSize;
+	private List<Batch> batches = new ArrayList<Batch>();
 
-    private List<Batch> batches = new ArrayList<Batch>();
+	public BatchAnalyser(String batchVariable, int batchSize)
+	{
+		this.batchVariable = batchVariable;
+		this.batchSize = batchSize;
+	}
 
-    public int getBatchesSize()
-    {
-        return batches.size();
-    }
+	public int getBatchesSize()
+	{
+		return batches.size();
+	}
 
-    private class Batch
-    {
-        public int number = -1;
-        public String[] values;
+	public int getBatchNum(Map<String, Object> map)
+	{
+		int batchNumber;
 
-        public int filledSize = 0;
+		String taskBatchVariableValue = (String) map.get(batchVariable);
 
-        public Batch(int num, int size)
-        {
-            this.number = num;
-            values = new String[size];
-        }
+		// first look, if variable value is already in existing batches
+		batchNumber = findInBatches(taskBatchVariableValue);
 
-        public void addValue(String value)
-        {
-            values[filledSize] = value;
-            filledSize++;
-        }
-    }
+		if (batchNumber == -1)
+		{
+			batchNumber = addToBatches(taskBatchVariableValue);
+		}
 
-    public BatchAnalyser(String batchVariable, int batchSize)
-    {
-        this.batchVariable = batchVariable;
-        this.batchSize = batchSize;
-    }
+		return batchNumber;
+	}
 
-    public int getBatchNum(Map<String, Object> map)
-    {
-        int batchNumber;
+	private int addToBatches(String taskBatchVariableValue)
+	{
+		// get the latest batch
+		boolean createNewBatch = false;
+		if (batches.size() > 0)
+		{
+			Batch batch = batches.get(batches.size() - 1);
 
-        String taskBatchVariableValue = (String) map.get(batchVariable);
+			if (batch.filledSize != batchSize)
+			{
+				batch.addValue(taskBatchVariableValue);
+				return batch.number;
+			}
+			else createNewBatch = true;
+		}
+		else createNewBatch = true;
 
-        //first look, if variable value is already in existing batches
-        batchNumber = findInBatches(taskBatchVariableValue);
+		if (createNewBatch)
+		{
+			// create new batch
+			Batch batch = new Batch(batches.size(), batchSize);
+			batches.add(batch);
+			batch.batches[0] = taskBatchVariableValue;
+			return batch.number;
+		}
+		return -1;
+	}
 
-        if(batchNumber == -1)
-        {
-            batchNumber = addToBatches(taskBatchVariableValue);
-        }
-
-        return batchNumber;
-    }
-
-    private int addToBatches(String taskBatchVariableValue)
-    {
-        //get the latest batch
-        boolean createNewBatch = false;
-        if(batches.size() > 0)
-        {
-            Batch batch = batches.get(batches.size() - 1);
-
-            if (batch.filledSize != batchSize)
-            {
-                batch.addValue(taskBatchVariableValue);
-                return batch.number;
-            }
-            else
-                createNewBatch = true;
-        }
-        else
-            createNewBatch = true;
-
-        if(createNewBatch)
-        {
-            //create new batch
-            Batch b = new Batch(batches.size(), batchSize);
-            batches.add(b);
-            b.values[0] = taskBatchVariableValue;
-            return b.number;
-        }
-        return -1;
-    }
-
-    private int findInBatches(String taskBatchVariableValue)
-    {
-        for(Batch batch : batches)
-        {
-            for(int i = 0; i < batch.filledSize; i++)
-            {
-                if(batch.values[i].equalsIgnoreCase(taskBatchVariableValue))
-                    return batch.number;
-            }
-        }
-        return -1;
-    }
+	private int findInBatches(String taskBatchVariableValue)
+	{
+		for (Batch batch : batches)
+		{
+			for (int i = 0; i < batch.filledSize; i++)
+			{
+				if (batch.batches[i].equalsIgnoreCase(taskBatchVariableValue)) return batch.number;
+			}
+		}
+		return -1;
+	}
 }
